Modifications singular

+++ matlab/+prost/private/config.hpp
@@ -3,8 +3,8 @@
 
 namespace matlab {
 
-typedef float real;
-  //typedef double real;
+// typedef float real;
+typedef double real;
 
 } // namespace matlab
